Stop users without passwords from being able to remove google auth

+++ src/controllers/multi_auth_controller.cr
@@ -64,13 +64,20 @@ class MultiAuthController < ApplicationController
 
   def unlink
     if auth_provider = AuthProvider.find(params[:id])
-      auth_provider.destroy
+      hashed_password =  auth_provider.user!.hashed_password
 
-      flash[:success] = "Unlinked #{auth_provider.provider.capitalize} provider"
-      redirect_to "/profile/edit"
+      if hashed_password.nil? || hashed_password.empty?
+        flash[:danger] = "A password must be set on the account before removing other authentication providers"
+        redirect_to "/profile/edit"
+      else
+        auth_provider.destroy
+
+        flash[:success] = "Unlinked #{auth_provider.provider.capitalize} provider"
+        redirect_to "/profile/edit"
+      end
     else
-      flash[:error] = "Can't find linked authentication provider"
-      redirect_to "/"
+      flash[:danger] = "Can't find linked authentication provider"
+      redirect_to "/profile/edit"
     end
   end
